Make the tests work

+++ clfsm/test-suspend/TestOnResume.machine/State_Done_Suspending_OnEntry.mm
@@ -1,2 +1,5 @@
-unlink(testfile);
-cerr << " *** Error: should be suspended!" << endl;
+if (ok) cerr << "OK.\n" << endl;
+else {
+  unlink(testfile);
+  cerr << " *** Error: should be suspended!" << endl;
+}
