+++ clfsm/test-suspend/TestOnResume.machine/State_SUSPENDED_OnExit.mm
@@ -1,3 +1,10 @@
-cerr << "Error: suspend state is not final" << endl;
-unlink(testfile);
+if (ok) ok = resume_at(machine_index()) != CLError;
+if (ok) cerr << "OK.\n" << endl;
+else {
+ cerr << "*** Error: resume_self() did not work." << endl;
+  unlink(testfile);
+  exit(EXIT_FAILURE);
+}
+
+// should never run twice
 ok = false;
